change artifact name to com.presagetech:smartspectra
