+++ internal-demo/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         </activity>
 
         <meta-data
-            android:name="com.presagetech.physiology.api_key"
+            android:name="com.presagetech.smartspectra.api_key"
             android:value="0NYqErApXtwP0EWbXajK5ojl7Y9jVcZ9rpgnoeji" />
     </application>
 </manifest>
