--- sdk/src/main/java/com/presagetech/physiology/Config.kt
@@ -1,5 +1,0 @@
-package com.presagetech.physiology
-
-object Config {
-    const val SAVE_JSONs = true
-}
