--- sdk/src/main/java/com/presagetech/physiology/core/result/ScreeningResult.kt
@@ -1,10 +1,0 @@
-package com.presagetech.physiology.core.result
-
-sealed class ScreeningResult {
-    data class Success(
-        val hrAverage: Double, // average heart rate
-        val rrAverage: Double, // average respiratory rate
-    ) : ScreeningResult()
-
-    data object Failed : ScreeningResult()
-}
