--- sdk/src/main/java/com/presagetech/physiology/core/result/SmartSpectraResultListener.kt
@@ -1,5 +1,0 @@
-package com.presagetech.physiology.core.result
-
-fun interface SmartSpectraResultListener {
-    fun onResult(result: ScreeningResult)
-}
