--- sdk/src/main/java/com/presagetech/physiology/network/HttpResponse.kt
@@ -1,7 +1,0 @@
-package com.presagetech.physiology.network
-
-class HttpResponse(
-    val responseCode: Int,
-    val responseHeaders: Map<String, List<String>>,
-    val responseBody: String?
-)
