+++ sdk/src/main/java/com/presagetech/smartspectra/Config.kt
@@ -1,0 +1,5 @@
+package com.presagetech.smartspectra
+
+object Config {
+    const val SAVE_JSONs = true
+}
