+++ sdk/src/main/java/com/presagetech/smartspectra/core/result/ScreeningResult.kt
@@ -1,0 +1,10 @@
+package com.presagetech.smartspectra.core.result
+
+sealed class ScreeningResult {
+    data class Success(
+        val hrAverage: Double, // average heart rate
+        val rrAverage: Double, // average respiratory rate
+    ) : ScreeningResult()
+
+    data object Failed : ScreeningResult()
+}
