+++ sdk/src/main/java/com/presagetech/smartspectra/core/result/SmartSpectraResultListener.kt
@@ -1,0 +1,5 @@
+package com.presagetech.smartspectra.core.result
+
+fun interface SmartSpectraResultListener {
+    fun onResult(result: ScreeningResult)
+}
