+++ sdk/src/main/java/com/presagetech/smartspectra/network/HttpResponse.kt
@@ -1,0 +1,7 @@
+package com.presagetech.smartspectra.network
+
+class HttpResponse(
+    val responseCode: Int,
+    val responseHeaders: Map<String, List<String>>,
+    val responseBody: String?
+)
